cpufreq: Underclock cpu cluster to 1689600 Khz
The underclock can be disabled by specifying "cpufreq.disable_underclock=1"
in the kernel command line, or by writing to disable_underclock at runtime.
Note that we need to underclock like this instead of outright removing the
higher frequencies because removing the higher frequencies outright causes
a significant decrease in performance for some reason.

Signed-off-by: Sultan Alsawaf <sultanxda@gmail.com>
[modified to msm8953]
Signed-off-by: Dede Dindin Qudsy <xtrymind@gmail.com>