cpufreq: stats: update uid data in acct_update_power
This eliminates the need to maintain separate alive_* and dead_*
arrays for each uid, as well as the need to loop over every task and
repeatedly take locks when printing out per-uid stats.

It also should facilitate replacing the use of single_open with seq
iterator because readers no longer need to take uid_lock or modify
uid_hashtable.

Test: Read /proc/uid_time_in_state, /proc/uid_cpupower/time_in_state
and /proc/uid/<uid>/time_in_state and check that values are
reasonable.
Test: Reading /proc/uid_cpupower/time_in_state takes <1ms
Bug: 71718257
Bug: 73328000
Signed-off-by: Connor O'Brien <connoro@google.com>

Change-Id: I607596efed00c3beccb468a0b3e25781646e7794
Signed-off-by: Dede Dindin Qudsy <xtrymind@gmail.com>